WIP: monitor produce events

+++ src/worktree/monitor/Message.zig
@@ -1,4 +1,4 @@
 pub const Message = union(enum) {
-    add: struct { path: []u8, id: usize },
-    remove: usize,
+    add: struct { path: []u8, id: u64 },
+    remove: u64,
 };
